// sapui5-rt-1.108.31/resources/~20240508083600~/sap/fe/core/converters/ManifestSettings.ts
export enum TemplateType {
	ListReport = "ListReport",
	ObjectPage = "ObjectPage",
	AnalyticalListPage = "AnalyticalListPage"
}

export enum ActionType {
	DataFieldForAction = "ForAction",
	DataFieldForIntentBasedNavigation = "ForNavigation",
	Default = "Default",
	Primary = "Primary",
	Secondary = "Secondary",
	SwitchToActiveObject = "SwitchToActiveObject",
	SwitchToDraftObject = "SwitchToDraftObject",
	DraftActions = "DraftActions",
	CollaborationAvatars = "CollaborationAvatars",
	DefaultApply = "DefaultApply",
	Menu = "Menu",
	ShowFormDetails = "ShowFormDetails"
}

export enum SelectionMode {
	Auto = "Auto",
	None = "None",
	Multi = "Multi",
	Single = "Single"
}

export enum VariantManagementType {
	Page = "Page",
	Control = "Control",
	None = "None"
}

export enum CreationMode {
	NewPage = "NewPage",
	Inline = "Inline",
	CreationRow = "CreationRow",
	InlineCreationRows = "InlineCreationRows",
	External = "External"
}

export enum VisualizationType {
	Table = "Table",
	Chart = "Chart"
}
// ...
export type ChartPersonalizationManifestSettings =
	| boolean
	| {
			sort: boolean;
			type: boolean;
			item: boolean;
	  };

export type ChartManifestConfiguration = {
	chartSettings: {
		personalization: ChartPersonalizationManifestSettings;
	};
};